requirements.txt
numpy
soundfile
pillow
kokoro
moviepy
google-genai
SentencePiece
python-dotenv
directory-tree
streamlit
GitPython